commit 5-24

+++ app/src/main/java/cn/service/bean/CouponInfo.java
@@ -89,7 +89,7 @@ public class CouponInfo {
 	}
 
 	public float getPrice() {
-		return price;
+		return (int)price;
 	}
 
 	public void setPrice(float price) {
